requirements.txt
requests>=2.31
tqdm>=4.66
pandas>=2.2
pymongo>=4.7
mysql-connector-python==8.2.0
python-dotenv==1.0.0